+++ termbox/src/Termbox/Internal/Key.hs
@@ -14,7 +14,7 @@ module Termbox.Internal.Key
   )
 where
 
-import qualified Termbox.Bindings
+import qualified Termbox.Bindings.Hs
 
 -- | A key event.
 --
@@ -119,63 +119,63 @@ pattern KeyCtrlI = KeyTab
 pattern KeyCtrlUnderscore :: Key
 pattern KeyCtrlUnderscore = KeyCtrlSlash
 
-parseKey :: Termbox.Bindings.Tb_key -> Key
+parseKey :: Termbox.Bindings.Hs.Tb_key -> Key
 parseKey = \case
-  Termbox.Bindings.TB_KEY_ARROW_DOWN -> KeyArrowDown
-  Termbox.Bindings.TB_KEY_ARROW_LEFT -> KeyArrowLeft
-  Termbox.Bindings.TB_KEY_ARROW_RIGHT -> KeyArrowRight
-  Termbox.Bindings.TB_KEY_ARROW_UP -> KeyArrowUp
-  Termbox.Bindings.TB_KEY_BACKSPACE -> KeyBackspace
-  Termbox.Bindings.TB_KEY_CTRL_TILDE -> KeyCtrlTilde
-  Termbox.Bindings.TB_KEY_CTRL_6 -> KeyCtrl6
-  Termbox.Bindings.TB_KEY_CTRL_8 -> KeyCtrl8
-  Termbox.Bindings.TB_KEY_CTRL_A -> KeyCtrlA
-  Termbox.Bindings.TB_KEY_CTRL_B -> KeyCtrlB
-  Termbox.Bindings.TB_KEY_CTRL_BACKSLASH -> KeyCtrlBackslash
-  Termbox.Bindings.TB_KEY_CTRL_C -> KeyCtrlC
-  Termbox.Bindings.TB_KEY_CTRL_D -> KeyCtrlD
-  Termbox.Bindings.TB_KEY_CTRL_E -> KeyCtrlE
-  Termbox.Bindings.TB_KEY_CTRL_F -> KeyCtrlF
-  Termbox.Bindings.TB_KEY_CTRL_G -> KeyCtrlG
-  Termbox.Bindings.TB_KEY_CTRL_H -> KeyCtrlBackspace
-  Termbox.Bindings.TB_KEY_CTRL_J -> KeyCtrlJ
-  Termbox.Bindings.TB_KEY_CTRL_K -> KeyCtrlK
-  Termbox.Bindings.TB_KEY_CTRL_L -> KeyCtrlL
-  Termbox.Bindings.TB_KEY_CTRL_N -> KeyCtrlN
-  Termbox.Bindings.TB_KEY_CTRL_O -> KeyCtrlO
-  Termbox.Bindings.TB_KEY_CTRL_P -> KeyCtrlP
-  Termbox.Bindings.TB_KEY_CTRL_Q -> KeyCtrlQ
-  Termbox.Bindings.TB_KEY_CTRL_R -> KeyCtrlR
-  Termbox.Bindings.TB_KEY_CTRL_RSQ_BRACKET -> KeyCtrlRsqBracket
-  Termbox.Bindings.TB_KEY_CTRL_S -> KeyCtrlS
-  Termbox.Bindings.TB_KEY_CTRL_SLASH -> KeyCtrlSlash
-  Termbox.Bindings.TB_KEY_CTRL_T -> KeyCtrlT
-  Termbox.Bindings.TB_KEY_CTRL_U -> KeyCtrlU
-  Termbox.Bindings.TB_KEY_CTRL_V -> KeyCtrlV
-  Termbox.Bindings.TB_KEY_CTRL_W -> KeyCtrlW
-  Termbox.Bindings.TB_KEY_CTRL_X -> KeyCtrlX
-  Termbox.Bindings.TB_KEY_CTRL_Y -> KeyCtrlY
-  Termbox.Bindings.TB_KEY_CTRL_Z -> KeyCtrlZ
-  Termbox.Bindings.TB_KEY_DELETE -> KeyDelete
-  Termbox.Bindings.TB_KEY_END -> KeyEnd
-  Termbox.Bindings.TB_KEY_ENTER -> KeyEnter
-  Termbox.Bindings.TB_KEY_ESC -> KeyEsc
-  Termbox.Bindings.TB_KEY_F1 -> KeyF1
-  Termbox.Bindings.TB_KEY_F10 -> KeyF10
-  Termbox.Bindings.TB_KEY_F11 -> KeyF11
-  Termbox.Bindings.TB_KEY_F12 -> KeyF12
-  Termbox.Bindings.TB_KEY_F2 -> KeyF2
-  Termbox.Bindings.TB_KEY_F3 -> KeyF3
-  Termbox.Bindings.TB_KEY_F4 -> KeyF4
-  Termbox.Bindings.TB_KEY_F5 -> KeyF5
-  Termbox.Bindings.TB_KEY_F6 -> KeyF6
-  Termbox.Bindings.TB_KEY_F7 -> KeyF7
-  Termbox.Bindings.TB_KEY_F8 -> KeyF8
-  Termbox.Bindings.TB_KEY_F9 -> KeyF9
-  Termbox.Bindings.TB_KEY_HOME -> KeyHome
-  Termbox.Bindings.TB_KEY_INSERT -> KeyInsert
-  Termbox.Bindings.TB_KEY_PGDN -> KeyPageDn
-  Termbox.Bindings.TB_KEY_PGUP -> KeyPageUp
-  Termbox.Bindings.TB_KEY_SPACE -> KeySpace
-  Termbox.Bindings.TB_KEY_TAB -> KeyTab
+  Termbox.Bindings.Hs.TB_KEY_ARROW_DOWN -> KeyArrowDown
+  Termbox.Bindings.Hs.TB_KEY_ARROW_LEFT -> KeyArrowLeft
+  Termbox.Bindings.Hs.TB_KEY_ARROW_RIGHT -> KeyArrowRight
+  Termbox.Bindings.Hs.TB_KEY_ARROW_UP -> KeyArrowUp
+  Termbox.Bindings.Hs.TB_KEY_BACKSPACE -> KeyBackspace
+  Termbox.Bindings.Hs.TB_KEY_CTRL_TILDE -> KeyCtrlTilde
+  Termbox.Bindings.Hs.TB_KEY_CTRL_6 -> KeyCtrl6
+  Termbox.Bindings.Hs.TB_KEY_CTRL_8 -> KeyCtrl8
+  Termbox.Bindings.Hs.TB_KEY_CTRL_A -> KeyCtrlA
+  Termbox.Bindings.Hs.TB_KEY_CTRL_B -> KeyCtrlB
+  Termbox.Bindings.Hs.TB_KEY_CTRL_BACKSLASH -> KeyCtrlBackslash
+  Termbox.Bindings.Hs.TB_KEY_CTRL_C -> KeyCtrlC
+  Termbox.Bindings.Hs.TB_KEY_CTRL_D -> KeyCtrlD
+  Termbox.Bindings.Hs.TB_KEY_CTRL_E -> KeyCtrlE
+  Termbox.Bindings.Hs.TB_KEY_CTRL_F -> KeyCtrlF
+  Termbox.Bindings.Hs.TB_KEY_CTRL_G -> KeyCtrlG
+  Termbox.Bindings.Hs.TB_KEY_CTRL_H -> KeyCtrlBackspace
+  Termbox.Bindings.Hs.TB_KEY_CTRL_J -> KeyCtrlJ
+  Termbox.Bindings.Hs.TB_KEY_CTRL_K -> KeyCtrlK
+  Termbox.Bindings.Hs.TB_KEY_CTRL_L -> KeyCtrlL
+  Termbox.Bindings.Hs.TB_KEY_CTRL_N -> KeyCtrlN
+  Termbox.Bindings.Hs.TB_KEY_CTRL_O -> KeyCtrlO
+  Termbox.Bindings.Hs.TB_KEY_CTRL_P -> KeyCtrlP
+  Termbox.Bindings.Hs.TB_KEY_CTRL_Q -> KeyCtrlQ
+  Termbox.Bindings.Hs.TB_KEY_CTRL_R -> KeyCtrlR
+  Termbox.Bindings.Hs.TB_KEY_CTRL_RSQ_BRACKET -> KeyCtrlRsqBracket
+  Termbox.Bindings.Hs.TB_KEY_CTRL_S -> KeyCtrlS
+  Termbox.Bindings.Hs.TB_KEY_CTRL_SLASH -> KeyCtrlSlash
+  Termbox.Bindings.Hs.TB_KEY_CTRL_T -> KeyCtrlT
+  Termbox.Bindings.Hs.TB_KEY_CTRL_U -> KeyCtrlU
+  Termbox.Bindings.Hs.TB_KEY_CTRL_V -> KeyCtrlV
+  Termbox.Bindings.Hs.TB_KEY_CTRL_W -> KeyCtrlW
+  Termbox.Bindings.Hs.TB_KEY_CTRL_X -> KeyCtrlX
+  Termbox.Bindings.Hs.TB_KEY_CTRL_Y -> KeyCtrlY
+  Termbox.Bindings.Hs.TB_KEY_CTRL_Z -> KeyCtrlZ
+  Termbox.Bindings.Hs.TB_KEY_DELETE -> KeyDelete
+  Termbox.Bindings.Hs.TB_KEY_END -> KeyEnd
+  Termbox.Bindings.Hs.TB_KEY_ENTER -> KeyEnter
+  Termbox.Bindings.Hs.TB_KEY_ESC -> KeyEsc
+  Termbox.Bindings.Hs.TB_KEY_F1 -> KeyF1
+  Termbox.Bindings.Hs.TB_KEY_F10 -> KeyF10
+  Termbox.Bindings.Hs.TB_KEY_F11 -> KeyF11
+  Termbox.Bindings.Hs.TB_KEY_F12 -> KeyF12
+  Termbox.Bindings.Hs.TB_KEY_F2 -> KeyF2
+  Termbox.Bindings.Hs.TB_KEY_F3 -> KeyF3
+  Termbox.Bindings.Hs.TB_KEY_F4 -> KeyF4
+  Termbox.Bindings.Hs.TB_KEY_F5 -> KeyF5
+  Termbox.Bindings.Hs.TB_KEY_F6 -> KeyF6
+  Termbox.Bindings.Hs.TB_KEY_F7 -> KeyF7
+  Termbox.Bindings.Hs.TB_KEY_F8 -> KeyF8
+  Termbox.Bindings.Hs.TB_KEY_F9 -> KeyF9
+  Termbox.Bindings.Hs.TB_KEY_HOME -> KeyHome
+  Termbox.Bindings.Hs.TB_KEY_INSERT -> KeyInsert
+  Termbox.Bindings.Hs.TB_KEY_PGDN -> KeyPageDn
+  Termbox.Bindings.Hs.TB_KEY_PGUP -> KeyPageUp
+  Termbox.Bindings.Hs.TB_KEY_SPACE -> KeySpace
+  Termbox.Bindings.Hs.TB_KEY_TAB -> KeyTab
   key -> error ("unknown key: " ++ show key)