+++ termbox/src/Termbox/Internal/Mouse.hs
@@ -11,11 +11,11 @@ module Termbox.Internal.Mouse
   )
 where
 
-import qualified Termbox.Bindings
+import qualified Termbox.Bindings.Hs
 
 -- | A mouse event.
 newtype Mouse
-  = Mouse Termbox.Bindings.Tb_key
+  = Mouse Termbox.Bindings.Hs.Tb_key
   deriving stock (Eq, Ord)
 
 instance Show Mouse where
@@ -28,21 +28,21 @@ instance Show Mouse where
     MouseWheelUp -> "MouseWheelUp"
 
 pattern MouseLeft :: Mouse
-pattern MouseLeft = Mouse Termbox.Bindings.TB_KEY_MOUSE_LEFT
+pattern MouseLeft = Mouse Termbox.Bindings.Hs.TB_KEY_MOUSE_LEFT
 
 pattern MouseMiddle :: Mouse
-pattern MouseMiddle = Mouse Termbox.Bindings.TB_KEY_MOUSE_MIDDLE
+pattern MouseMiddle = Mouse Termbox.Bindings.Hs.TB_KEY_MOUSE_MIDDLE
 
 pattern MouseRelease :: Mouse
-pattern MouseRelease = Mouse Termbox.Bindings.TB_KEY_MOUSE_RELEASE
+pattern MouseRelease = Mouse Termbox.Bindings.Hs.TB_KEY_MOUSE_RELEASE
 
 pattern MouseRight :: Mouse
-pattern MouseRight = Mouse Termbox.Bindings.TB_KEY_MOUSE_RIGHT
+pattern MouseRight = Mouse Termbox.Bindings.Hs.TB_KEY_MOUSE_RIGHT
 
 pattern MouseWheelDown :: Mouse
-pattern MouseWheelDown = Mouse Termbox.Bindings.TB_KEY_MOUSE_WHEEL_DOWN
+pattern MouseWheelDown = Mouse Termbox.Bindings.Hs.TB_KEY_MOUSE_WHEEL_DOWN
 
 pattern MouseWheelUp :: Mouse
-pattern MouseWheelUp = Mouse Termbox.Bindings.TB_KEY_MOUSE_WHEEL_UP
+pattern MouseWheelUp = Mouse Termbox.Bindings.Hs.TB_KEY_MOUSE_WHEEL_UP
 
 {-# COMPLETE MouseLeft, MouseMiddle, MouseRelease, MouseRight, MouseWheelDown, MouseWheelUp #-}