+++ termbox/src/Termbox/Internal/Scene.hs
@@ -7,7 +7,7 @@ module Termbox.Internal.Scene
   )
 where
 
-import qualified Termbox.Bindings
+import qualified Termbox.Bindings.Hs
 import Termbox.Internal.Cell (Cell, drawCell)
 import Termbox.Internal.Color (Color (Color))
 import Termbox.Internal.Pos (Pos (..))
@@ -19,8 +19,8 @@ import Termbox.Internal.Pos (Pos (..))
 -- * Set the cursor position with 'cursor'.
 -- * Combine scenes together with @<>@.
 data Scene = Scene
-  { sceneFill :: Maybe Termbox.Bindings.Tb_color,
-    sceneDraw :: Termbox.Bindings.Tb_color -> IO ()
+  { sceneFill :: Maybe Termbox.Bindings.Hs.Tb_color,
+    sceneDraw :: Termbox.Bindings.Hs.Tb_color -> IO ()
   }
 
 instance Monoid Scene where
@@ -48,12 +48,12 @@ drawScene :: Scene -> IO ()
 drawScene Scene {sceneFill, sceneDraw} = do
   let background =
         case sceneFill of
-          Nothing -> Termbox.Bindings.TB_DEFAULT
+          Nothing -> Termbox.Bindings.Hs.TB_DEFAULT
           Just color -> color
-  Termbox.Bindings.tb_set_clear_attributes Termbox.Bindings.TB_DEFAULT background
-  Termbox.Bindings.tb_clear
+  Termbox.Bindings.Hs.tb_set_clear_attributes Termbox.Bindings.Hs.TB_DEFAULT background
+  Termbox.Bindings.Hs.tb_clear
   sceneDraw background
-  Termbox.Bindings.tb_present
+  Termbox.Bindings.Hs.tb_present
 
 -- | Set the background fill color.
 fill :: Color -> Scene
@@ -68,4 +68,4 @@ cell Pos {col, row} img =
 -- | Set the cursor position.
 cursor :: Pos -> Scene
 cursor Pos {col, row} =
-  mempty {sceneDraw = \_ -> Termbox.Bindings.tb_set_cursor (Just (col, row))}
+  mempty {sceneDraw = \_ -> Termbox.Bindings.Hs.tb_set_cursor (Just (col, row))}